Converges
The U0, U1 were messed up before because they were continously being
updated when they should have been fixed throughout the whole guass
seidel function. The force was made fixed as well because it was being
continouosly being updated before, which is incorrect because U1 and U0
are supposed to be constant.